% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceLmer.R
\name{ExtractRandomEffect}
\alias{ExtractRandomEffect}
\title{Extract random effects from lmer with multiple imputation datasets using mice}
\usage{
ExtractRandomEffect(
  n_levels = 2,
  imputed_datasets,
  lmer_formula,
  group_v1,
  group_v2 = NULL
)
}
\arguments{
\item{n_levels}{level of multilevel models}

\item{imputed_datasets}{mice output.}

\item{lmer_formula}{Formula for lmer. The class of this parameter has to be character.}

\item{group_v1}{Clustering variable. If you applying three-level multilevel model, group_v1 has to be a level2 variable.}

\item{group_v2}{Clustering variable at three-level multilevel models. group_v2 has to be a level3 variable. If you applying two-level multilevel model, group_v2 has to be NA.}
}
\value{
data.frame including random effects at lmer results with mice.
}
\description{
\code{ExtractRandomEffect} Return dataframe including random effects at lmer results with mice
}
