% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceLmer.R
\name{GetProportionOfRandomEffect}
\alias{GetProportionOfRandomEffect}
\title{Combine random effect parameters calculated from multiply imputed datasets based on Rubin's rule}
\usage{
GetProportionOfRandomEffect(data)
}
\arguments{
\item{data}{output of ExtractRandomEffect function.}
}
\value{
dataframe including random effect parameters (variance and standard deviation) and proportions of the total variance
}
\description{
\code{GetProportionOfRandomEffect} Return random effect parameters (variance and standard deviation) and proportions of the total variance
}
