% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceLmer.R
\name{MiceLmer}
\alias{MiceLmer}
\alias{ExtractRandomEffect}
\alias{PlotRandomEffectDistribution}
\alias{GetProportionOfRandomEffect}
\title{Extract random effects from lmer with multiple imputation datasets using mice}
\usage{
ExtractRandomEffect(imputed_datasets, lmer_formula, group_v1, group_v2 = NA)

PlotRandomEffectDistribution(data, group_v1, group_v2 = NA)

GetProportionOfRandomEffect(data)
}
\arguments{
\item{imputed_datasets}{mice output.}

\item{lmer_formula}{Formula for lmer. The class of this parameter has to be character.}

\item{group_v1}{Clustering variable. If you applying three-level multilevel model, group_v1 has to be a level2 variable.}

\item{group_v2}{Clustering variable at three-level multilevel models. group_v2 has to be a level3 variable. If you applying two-level multilevel model, group_v2 has to be NA.}

\item{data}{output of ExtractRandomEffect function.}
}
\value{
data.frame including random effects at lmer results with mice.

distributions of random effects

dataframe including random effect parameters (variance and standard deviation) and proportions of the total variance
}
\description{
\code{ExtractRandomEffect} Return dataframe including random effects at lmer results with mice

\code{PlotRandomEffectDistribution} To confirm the parameters are following normal distributions (Rubin's rule), this function shows distributions of random effects parameters

\code{GetProportionOfRandomEffect} Return random effect parameters (variance and standard deviation) and proportions of the total variance
}
