% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiceLmer.R
\name{PlotRandomEffectDistribution}
\alias{PlotRandomEffectDistribution}
\title{Plot distributions of random effects}
\usage{
PlotRandomEffectDistribution(data, group_v1, group_v2 = NA)
}
\arguments{
\item{data}{output of ExtractRandomEffect function.}

\item{group_v1}{Clustering variable. If you applying three-level multilevel model, group_v1 has to be a level2 variable.}

\item{group_v2}{Clustering variable at three-level multilevel models. group_v2 has to be a level3 variable. If you applying two-level multilevel model, group_v2 has to be NA.}
}
\value{
distributions of random effects
}
\description{
\code{PlotRandomEffectDistribution} To confirm the parameters are following normal distributions (Rubin's rule), this function shows distributions of random effects parameters
}
